backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
psycopg2-binary==2.8.6
python-dotenv==0.17.1
django-cpf==0.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
git+https://github.com/jadsonlucio/pyGRU.git@main
